+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/WProgressBar_Test.java
@@ -1,6 +1,6 @@
 package com.github.bordertech.wcomponents;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 import org.junit.Test;
 
 /**
@@ -12,100 +12,102 @@ import org.junit.Test;
 public class WProgressBar_Test extends AbstractWComponentTestCase {
 
 	@Test
+	public void testDefaultConstructor() {
+		WProgressBar bar = new WProgressBar();
+		Assert.assertEquals(bar.getProgressBarType(), WProgressBar.DEFAULT_TYPE);
+		Assert.assertEquals(0, bar.getMax());
+		Assert.assertEquals(0, bar.getValue());
+	}
+
+	@Test
+	public void testConstructorWithType() {
+		WProgressBar bar = new WProgressBar(WProgressBar.ProgressBarType.SMALL);
+		Assert.assertEquals(bar.getProgressBarType(), WProgressBar.ProgressBarType.SMALL);
+		Assert.assertEquals(0, bar.getMax());
+		Assert.assertEquals(0, bar.getValue());
+		bar = new WProgressBar(WProgressBar.ProgressBarType.NORMAL);
+		Assert.assertEquals(bar.getProgressBarType(), WProgressBar.ProgressBarType.NORMAL);
+		Assert.assertEquals(0, bar.getMax());
+		Assert.assertEquals(0, bar.getValue());
+	}
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testConstructorWithNullType() {
+		WProgressBar progressBar = new WProgressBar(null);
+	}
+
+	@Test
+	public void testContructorWithMax() {
+		int max = 10;
+		WProgressBar bar = new WProgressBar(max);
+		Assert.assertEquals(max, bar.getMax());
+	}
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testNegativeMaxInConstructor() {
+		WProgressBar progressBar = new WProgressBar(-1);
+	}
+
+	@Test
+	public void testConstructorTypeAndMax() {
+		int max = 10;
+		WProgressBar.ProgressBarType type = WProgressBar.ProgressBarType.SMALL;
+		WProgressBar bar = new WProgressBar(type, max);
+		Assert.assertEquals(type, bar.getProgressBarType());
+		Assert.assertEquals(max, bar.getMax());
+	}
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testConstructorWithTypeAndMaxWithNullType() {
+		WProgressBar progressBar = new WProgressBar(null, 10);
+	}
+
+	@Test(expected = IllegalArgumentException.class)
+	public void testConstructorWithTypeAndMaxWithNegativeMax() {
+		WProgressBar progressBar = new WProgressBar(WProgressBar.ProgressBarType.SMALL, -1);
+	}
+
+	@Test
 	public void testTypeAccessors() {
-		final WProgressBar.ProgressBarType type1 = WProgressBar.ProgressBarType.NORMAL;
-		final WProgressBar.ProgressBarType type2 = WProgressBar.ProgressBarType.SMALL;
-		final WProgressBar.ProgressBarType type3 = WProgressBar.ProgressBarType.NORMAL;
-
-		WProgressBar progressBar = new WProgressBar(type1, WProgressBar.UnitType.FRACTION);
-		Assert.assertEquals("Incorrect progressBar type", type1, progressBar.getProgressBarType());
-
-		progressBar.setProgressBarType(type2);
-		Assert.assertEquals("Incorrect progressBar type after setProgressBarType", type2,
-				progressBar.getProgressBarType());
-
-		progressBar.setLocked(true);
-		setActiveContext(createUIContext());
-		progressBar.setProgressBarType(type3);
-
-		Assert.assertEquals("Incorrect progressBar type after session setProgressBarType", type3,
-				progressBar.getProgressBarType());
-
-		resetContext();
-		Assert.assertEquals("Incorrect default progressBar type after session setProgressBarType",
-				type2, progressBar.getProgressBarType());
+		assertAccessorsCorrect(new WProgressBar(), "progressBarType", WProgressBar.DEFAULT_TYPE, WProgressBar.ProgressBarType.SMALL,
+				WProgressBar.ProgressBarType.NORMAL);
 	}
 
 	@Test
-	public void testUnitTypeAccessors() {
-		final WProgressBar.UnitType unitType1 = WProgressBar.UnitType.FRACTION;
-		final WProgressBar.UnitType unitType2 = WProgressBar.UnitType.PERCENTAGE;
-
-		WProgressBar progressBar = new WProgressBar(WProgressBar.ProgressBarType.NORMAL, unitType1);
-		Assert.assertEquals("Incorrect progressBar unit type", unitType1, progressBar.getUnitType());
-
-		progressBar.setUnitType(unitType2);
-		Assert.assertEquals("Incorrect progressBar unit type after setUnitType", unitType2,
-				progressBar.getUnitType());
-
-		progressBar.setLocked(true);
-		setActiveContext(createUIContext());
-		progressBar.setUnitType(unitType1);
-
-		Assert.assertEquals("Incorrect progressBar unit type after session setUnitType", unitType1,
-				progressBar.getUnitType());
-
-		resetContext();
-		Assert.assertEquals("Incorrect default progressBar unit type after session setUnitType",
-				unitType2, progressBar.getUnitType());
-	}
-
-	@Test
-	public void testTextAccessors() {
-		WProgressBar progressBar = new WProgressBar();
-		String defaultText = "WProgressBar_Test.testGetText.default";
-		String userText = "WProgressBar_Test.testGetText.user";
-		progressBar.setText(defaultText);
-
-		Assert.assertEquals("Incorrect default text", defaultText, progressBar.getText());
-
-		// Set test for a users session
-		progressBar.setLocked(true);
-		setActiveContext(createUIContext());
-		progressBar.setText(userText);
-		Assert.assertEquals("Modified session should have session text", userText, progressBar.
-				getText());
-
-		resetContext();
-		Assert.assertEquals("Other session should have default text", defaultText, progressBar.
-				getText());
-
-		//Test nulls
-		progressBar.setText("");
-		Assert.assertEquals("text should be empty string", "", progressBar.getText());
-		progressBar.setText(null);
-		Assert.assertNull("text should be null", progressBar.getText());
+	public void testNullTypeInSetter() {
+		WProgressBar progressBar = new WProgressBar(WProgressBar.ProgressBarType.SMALL);
+		progressBar.setProgressBarType(null);
+		Assert.assertEquals("Expected type to be set to NORMAL", WProgressBar.DEFAULT_TYPE, progressBar.getProgressBarType());
 	}
 
 	@Test
 	public void testMaxAccessors() {
-		int max1 = 121;
-		int max2 = 122;
-		int max3 = 123;
+		assertAccessorsCorrect(new WProgressBar(), "max", 0, 10, 20);
+	}
 
-		WProgressBar progressBar = new WProgressBar(max1);
-		Assert.assertEquals("Max value accessors incorrect", max1, progressBar.getMax());
+	@Test(expected = IllegalArgumentException.class)
+	public void testNegativeMaxInSetter() {
+		WProgressBar progressBar = new WProgressBar();
+		progressBar.setMax(-1);
+	}
 
-		progressBar.setMax(max2);
-		Assert.assertEquals("Incorrect max value after call to setMax(int)", max2, progressBar.
-				getMax());
+	@Test
+	public void testValueAccessors() {
+		int max = 10;
+		assertAccessorsCorrect(new WProgressBar(max), "value", 0, (max - 2), (max - 1));
+	}
 
-		progressBar.setLocked(true);
-		setActiveContext(createUIContext());
-		progressBar.setMax(max3);
-		Assert.assertEquals("Incorrect max value for changed context", max3, progressBar.getMax());
+	@Test(expected = IllegalArgumentException.class)
+	public void testNegativeValue() {
+		WProgressBar progressBar = new WProgressBar();
+		progressBar.setValue(-1);
+	}
 
-		resetContext();
-		Assert.assertEquals("Default max value should not have changed", max2, progressBar.getMax());
+	@Test
+	public void testSetValueGreaterThanMaxGetsMax() {
+		int max = 10;
+		WProgressBar bar = new WProgressBar(max);
+		bar.setValue(max + 1);
+		Assert.assertEquals(max, bar.getValue());
 	}
 }